backend/requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-google-genai
langchain-core
pydantic
python-dotenv
scikit-learn
pandas
numpy
httpx
joblib
pytest
karrio
websockets
sqlalchemy
psycopg2-binary
redis
